refactor: kebab-case for updateDeps file
